Add onstart/onstop config option
The onstart and onstop options can be set in the configuration file to run a
command when the audio playback is about to begin or has ended.

The full path for the command has to be specified, e.g. '/usr/bin/logger'.

Executable scripts must begin with a shebang, e.g. '#!/bin/sh'.

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>