Add onvolumechange
Add an onvolumechange config option. This option can be set in the configuration
file to run a command or script when the volume is changed. The command or
script will be called with the volume as the first argument.

A use case for this might be to display the new volume level on a LCD display
connected to the media renderer.

Note that the full path for the command or script has to be specified, e.g.
'/usr/bin/logger'. Executable shell scripts must begin with a shebang, e.g.
'#!/bin/sh'.