Add onplay option
The onplay option is the counterpart to onstop. It can be set in the
configuration file to run a command or shell script when the audio playback
has began.

Note, that onplay is similiar to onstart, but differs in the way, that onstart
is called before playback has began while onplay is called when playback has
started.

The full path for the command has to be specified, e.g. '/usr/bin/logger'.

Executable scripts must begin with a shebang, e.g. '#!/bin/sh'.