Fix unsafe header path when cross-compiling
Building upmpdcli aborts because of an unsafe header inclusion of
`/usr/include/libupnpp`. Using `{prefix}` to get the header path is wrong for
cross-compilation. Instead of hard-coding the flags for libupnpp use pkg-config
to get the correct ones.

Also expand the LIBS by upnpp_LIBS to set the necessary libraries for
linking with libupnpp. This fixes static build issues.

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>