Fix running onplay after starting upmpdcli
If upmpdcli is started and MPD is already playing, the "onplay" command
is not run as the configuration files is not parsed yet.

Fix this by checking the state after parsing the configuration.
Additionally, a check for the state `MPDS_UNK` is necessary to run the
"onplay" command as upmpdcli might be just started.

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>