scctl: add -i/-I option to set source index
The `-i` option can be used to set the source by index, whereas the `-I`
option can be used to set the source by name for the OpenHome product service
of a UPnP renderer.

This is handy to enable/disable the Songcast Sender or any other
external source, e.g. use `scctl -I <device name> PL-to-Songcast` to enable the
Songcast Sender Service.

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>