Add scnosongcastsource config option
Add a "scnosongcastsource" config option to disable adding the
"PL-to-Songcast" and "RD-to-Songcast" as sources to the SourceXML.

Needs to be set to "1" to activate the option. The default behavior is
not changed. If the option is not set, the sources are added to
SourceXML.

Note, that setting this option will not allow you to use the Songcast
Tool in Upplay for the renderer, so you better know what you are doing.